requirements.txt
alembic==1.10.3
amqp==5.1.1
aniso8601==8.1.0
anytree==2.8.0
apispec==4.0.0
apispec-webframeworks==0.5.2
appdirs==1.4.4
argh==0.28.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-timeout==4.0.2
attrs==20.3.0
Authlib==1.3.1
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
black==24.3.0
blinker==1.8.2
cached-property==1.5.2
cachetools==5.3.0
celery==5.4.0
certifi==2024.7.4
cffi==1.14.4
cfgv==3.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colour==0.1.5
coverage==7.3.2
cryptography==42.0.4
decorator==5.1.0
dependency-injector==4.48.2
Deprecated==1.2.18
distlib==0.3.6
dnspython==2.6.1
docopt==0.6.2
dominate==2.6.0
email_validator==2.2.0
exceptiongroup==1.1.1
filelock==3.11.0
flake8==6.1.0
Flask==3.0.3
Flask-Admin==1.6.1
flask-apispec==0.11.4
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-Dance==6.2.0
Flask-gzip==0.2
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-QRcode==3.1.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
GeoAlchemy2==0.13.1
googlemaps==4.10.0
greenlet==3.0.3
gunicorn==22.0.0
icalendar==5.0.11
identify==1.5.10
idna==3.7
importlib-metadata==6.3.0
importlib_resources==6.4.0
iniconfig==1.1.1
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.4
johnnydep==1.18.0
jsonschema==3.2.0
kombu==5.4.0
limits==5.1.0
Mako==1.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
ordered-set==4.1.0
oyaml==1.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
pilkit==2.0
pillow==10.4.0
pipdeptree==2.7.0
pkginfo==1.9.6
platformdirs==3.1.0
pluggy==1.5.0
pre-commit==3.8.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pycparser==2.20
pyflakes==3.1.0
Pygments==2.19.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==8.3.4
pytest-celery==0.0.0
pytest-cov==4.0.0
pytest-datadir==1.4.1
pytest-mock==3.10.0
pytest-split==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
python-editor==1.0.4
pytoolconfig==1.2.5
pytz==2023.3
PyYAML==6.0.1
qrcode==6.1
recurring-ical-events==2.1.0
redis==4.5.4
regex==2023.3.23
requests==2.32.0
requests-mock==1.10.0
requests-oauthlib==1.3.0
rich==13.9.4
rope==1.7.0
setuptools==72.1.0
six==1.16.0
soupsieve==2.1
speaklater==1.3
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.0
structlog==23.1.0
swagger-spec-validator==3.0.3
tabulate==0.9.0
TatSu==5.8.3
toml==0.10.2
tomli==2.0.1
typed_ast==1.5.4
typing_extensions==4.6.0
tzdata==2024.1
urllib3==1.26.19
URLObject==2.4.3
validators==0.20.0
vine==5.1.0
virtualenv==20.21.0
visitor==0.1.3
wcwidth==0.2.6
webargs==7.0.1
Werkzeug==3.0.3
wheel==0.44.0
wimpy==0.6
wrapt==1.17.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
x-wr-timezone==0.0.7
yarg==0.1.9
zipp==3.19.1
